Fix secrets archive/deploy scripts

+++ scripts/archive-secrets.sh
@@ -45,15 +45,27 @@ echo-log info 'exporting GPG data.'
 mkdir ${working_dir}/${archive_dir}/gpg
 gpg --export -a --output ${working_dir}/${archive_dir}/gpg/public-keys.asc
 gpg --export-secret-keys --output ${working_dir}/${archive_dir}/gpg/secret-keys.asc
-chmod 600 ${working_dir}/${archive_dir}/gpg/secret-keys.asc
-gpg --export-ownertrust --output ${working_dir}/${archive_dir}/gpg/ownertrust
-chmod 600 ${working_dir}/${archive_dir}/gpg/ownertrust
-echo-log info 'exported GPG data.'
+if [ -f ${working_dir}/${archive_dir}/gpg/secret-keys.asc ] ; then
+  chmod 600 ${working_dir}/${archive_dir}/gpg/secret-keys.asc
+fi
+if gpg --export-ownertrust --output ${working_dir}/${archive_dir}/gpg/ownertrust ; then
+  chmod 600 ${working_dir}/${archive_dir}/gpg/ownertrust
+fi
+if [ $( ls -1 ${working_dir}/${archive_dir}/gpg | wc -l) -gt 0 ]; then
+  echo-log info 'exported GPG data.'
+else
+  echo-log warn 'no data to export, archive does not contain GPG data.'
+  rmdir ${working_dir}/${archive_dir}/gpg
+fi
 
 echo-log info 'copying Git files.'
-mkdir ${working_dir}/${archive_dir}/git
-cp -pr ~/.config/git/config ${working_dir}/${archive_dir}/git
-echo-log info 'copied Git files.'
+if [ -f ~/.config/git/config ] ;then
+  mkdir ${working_dir}/${archive_dir}/git
+  cp -pr ~/.config/git/config ${working_dir}/${archive_dir}/git
+  echo-log info 'copied Git files.'
+else
+  echo-log warn 'Git files not found, skipped copying.'
+fi
 
 echo-log info 'copying AWS CLI files.'
 if [ -f ~/.aws/config ]; then
@@ -68,7 +80,7 @@ echo-log ok 'copied secret files for archive.'
 
 
 echo-log info 'archive the following files.'
-tree ${working_dir}/${archive_dir}
+find ${working_dir}/${archive_dir}
 
 echo-log info 'archiving secret files.'
 tar -C ${working_dir} -zcvf ${working_dir}/${archive_dir}.tar.gz ${archive_dir}
@@ -80,6 +92,6 @@ rm -rf ${working_dir}/${archive_dir}
 echo-log info 'deleted copied secret files.'
 
 
-echo-log ok "secret files archived to ${working_dir}/${archive_dir}"
+echo-log ok "secret files archived to ${working_dir}/${archive_dir}.tar.gz"
 echo-log notice 'do not forget to delete it after use.'
